requirement.txt
APScheduler==3.6.3
asn1crypto==0.24.0
astroid==2.3.3
autopep8==1.5
blessed==1.15.0
boto3==1.9.253
botocore==1.12.253
bottle==0.12.18
CacheControl==0.12.5
cached-property==1.5.1
cachetools==3.1.1
cement==2.8.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
codename==1.1
colorama==0.3.9
cryptography==2.7
dateparser==0.7.2
decorator==4.4.0
deprecation==2.0.7
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
firebase-admin==3.0.0
Flask==1.1.1
Flask-Cache==0.13.1
Flask-Compress==1.4.0
Flask-Session==0.3.1
Flask-SocketIO==4.2.1
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
fpdf==1.7.2
future==0.16.0
gcloud==0.17.0
gevent==1.4.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.24.1
gunicorn==19.3.0
httplib2==0.13.1
idna==2.7
importlib-metadata==0.23
inexactsearch==1.0.2
isort==4.3.21
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
jsonobject==0.7.1
jsonpickle==0.9.6
jsonschema==2.6.0
jws==0.1.3
lazy-object-proxy==1.4.3
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.1
ndg-httpsclient==0.5.1
numpy==1.17.0
oauth2client==3.0.0
oauthlib==3.1.0
packaging==19.2
pandas==0.25.0
parso==0.5.1
passlib==1.7.1
pathspec==0.5.9
pipreqs==0.4.9
plivo==4.3.2
pluggy==0.13.0
protobuf==3.9.1
pusher==2.1.4
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.4.3
pygsheets==2.0.2
PyJWT==1.7.1
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-engineio==3.9.3
python-jsonrpc-server==0.2.0
python-jwt==2.0.1
python-language-server==0.28.3
python-shorturl==0.0.4
python-socketio==4.3.1
pytracking==0.2.0
pytz==2019.2
PyYAML==3.13
redis==3.3.8
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.7.0
rsa==4.0
s3transfer==0.2.1
semantic-version==2.5.0
silpa-common==0.3
six==1.14.0
soundex==1.1.3
squareconnect==2.20190724.1
squareup==3.20190814.0
stripe==2.42.0
taxjar==1.8.0
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.4.1
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.24.3
virtualenv==16.7.4
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
xlwt==1.3.0
yarg==0.1.9
zipp==0.6.0
